+++ src/net/bitpot/railways/gui/RoutesTable.java
@@ -1,15 +1,15 @@
 package net.bitpot.railways.gui;
 
-import com.intellij.ide.CopyProvider;
 import com.intellij.openapi.actionSystem.*;
 import com.intellij.openapi.diagnostic.Logger;
-import com.intellij.openapi.ide.CopyPasteManager;
 import com.intellij.ui.PopupHandler;
 import com.intellij.ui.table.JBTable;
+import com.intellij.util.ArrayUtil;
 import net.bitpot.railways.models.Route;
 import net.bitpot.railways.models.RouteTableModel;
 import org.jetbrains.annotations.NonNls;
 import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
 
 import javax.swing.*;
 import javax.swing.event.ListSelectionEvent;
@@ -24,13 +24,12 @@ import java.awt.event.MouseEvent;
 
 /**
  * We should add some behavior to default JTable to be able copy custom data.
- * To add ability to copy data we should follow next steps:
- * 1. Our component (JTable) should implement DataProvider interface. In this case when action is invoked
- * in context of this component, we will be able to pass out own data object, i.e. CopyProvider.
- * 2. Implement CopyProvider interface (it also could be done in private class) which is responsible for
- * changing Copy action status and copying custom information.
+ *
+ * Our component (JTable) should implement DataProvider interface. In this
+ * case when action is invoked in context of this component, we will be able to
+ * pass our own data to the action handler.
  */
-public class RoutesTable extends JBTable implements CopyProvider, DataProvider {
+public class RoutesTable extends JBTable implements DataProvider {
     @SuppressWarnings("unused")
     private static Logger log = Logger.getInstance(RoutesTable.class.getName());
 
@@ -117,48 +116,44 @@ public class RoutesTable extends JBTable implements CopyProvider, DataProvider {
      * @param dataId Data key
      * @return Data object.
      */
+    @Nullable
     @Override
     public Object getData(@NonNls String dataId) {
         // Good example of usage is in com.intellij.openapi.editor.impl.EditorComponentImpl (see getData method)
 
-        // We have custom copy provider, so return it when requested.
-        if (PlatformDataKeys.COPY_PROVIDER.is(dataId))
-            return this;
+        if (PlatformDataKeys.SELECTED_ITEMS.is(dataId))
+            return getSelectedRoutes();
+
+        if (PlatformDataKeys.SELECTED_ITEM.is(dataId))
+            return getSelectedRoute();
 
         return null;
     }
 
 
-    /**
-     * Performs copy operation of the first selected route in the table.
-     *
-     * @param dataContext Data context which action was invoked in.
-     */
-    @Override
-    public void performCopy(@NotNull DataContext dataContext) {
-        if (getSelectedRow() < 0)
-            return;
+    @Nullable
+    private Route getSelectedRoute() {
+        int selectedId = convertRowIndexToModel(getSelectedRow());
+        if (selectedId < 0)
+            return null;
 
-        Route route = ((RouteTableModel) getModel()).getRoute(convertRowIndexToModel(getSelectedRow()));
-        CopyPasteManager.getInstance().setContents(new StringSelection(route.getRouteName()));
+        return ((RouteTableModel) getModel()).getRoute(selectedId);
     }
 
 
-    /**
-     * Specifies conditions when copy action is enabled or disabled in routes popup menu.
-     *
-     * @param dataContext Data context.
-     * @return True or false.
-     */
-    @Override
-    public boolean isCopyEnabled(@NotNull DataContext dataContext) {
-        return getSelectedRowCount() > 0;
-    }
+    @NotNull
+    private Route[] getSelectedRoutes() {
+        RouteTableModel model = (RouteTableModel) getModel();
+        int[] selectedRows = getSelectedRows();
 
+        Route[] selectedRoutes = new Route[selectedRows.length];
 
-    @Override
-    public boolean isCopyVisible(@NotNull DataContext dataContext) {
-        return true;
+        for(int i = 0; i < selectedRows.length; i++) {
+            selectedRoutes[i] =
+                    model.getRoute(convertRowIndexToModel(selectedRows[i]));
+        }
+
+        return selectedRoutes;
     }
 
 
@@ -175,22 +170,34 @@ public class RoutesTable extends JBTable implements CopyProvider, DataProvider {
     private class MyPopupHandler extends PopupHandler {
 
         @Override
-        public void mouseReleased(MouseEvent e) {
-            // Here we override default behavior of mouseReleased to add implementation of
-            // row selection by right-click before popup is shown.
-            if (e.getButton() == MouseEvent.BUTTON3) {
-                int r = rowAtPoint(e.getPoint());
-                if (r >= 0 && r < getRowCount()) {
-                    setRowSelectionInterval(r, r);
-                } else {
-                    clearSelection();
-                }
-            }
+        public void mousePressed(MouseEvent e) {
+            handleRightClick(e);
+            super.mousePressed(e);
+        }
 
-            // And finally we can show handle popup
+        @Override
+        public void mouseReleased(MouseEvent e) {
+            handleRightClick(e);
             super.mouseReleased(e);
         }
 
+        private void handleRightClick(MouseEvent e) {
+            if (!SwingUtilities.isRightMouseButton(e) || !e.isPopupTrigger())
+                return;
+
+            // Before showing the popup, we should update selection properly:
+            //  * if right-clicked on existing selection - do nothing.
+            //  * if clicked outside current selection - clear it and select
+            //    only item that was clicked.
+            int clickedRowIndex = rowAtPoint(e.getPoint());
+            boolean isSelectionClicked =
+                    ArrayUtil.indexOf(getSelectedRows(), clickedRowIndex) >= 0;
+
+            if (!isSelectionClicked && clickedRowIndex >= 0 &&
+                    clickedRowIndex < getRowCount()) {
+                setRowSelectionInterval(clickedRowIndex, clickedRowIndex);
+            }
+        }
 
         @Override
         public void invokePopup(Component comp, int x, int y) {