+++ src/net/bitpot/railways/gui/StringFormatter.java
@@ -1,0 +1,5 @@
+package net.bitpot.railways.gui;
+
+public interface StringFormatter {
+    String format(String str);
+}