+++ CHANGES.md
@@ -1,6 +1,13 @@
 Railways - Change Notes
 =======================
 
+### 0.8.4 (2016-01-30)
+
+* Add ability to copy route path with http method
+* Add ability to copy data from multiple selected rows
+* Make right-click behavior more intuitive for the routes table
+
+
 ### 0.8.3 (2015-05-16)
 
 * Optimize performance of route action availability check.